Removed the fieldmap file, moved limelighthelpers to a library folder
